requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2022.6.15
charset-normalizer==2.1.0
cloudinary==1.29.0
Django==4.0.5
django-cloudinary-storage==0.3.0
django-cors-headers==3.13.0
django-jazzmin==2.5.0
djangorestframework==3.13.1
idna==3.3
Pillow==9.2.0
psycopg2==2.9.3
PyJWT==2.4.0
python-magic==0.4.27
pytz==2022.1
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.9
